requirements.txt
Django==3.1.0
djangorestframework==3.11.0
django-cors-headers==3.5.0
drf-yasg==1.20.0
